apis/ai-3d-model-api/requirements.txt
fastapi
uvicorn
python-dotenv
replicate
requests
python-multipart
pydantic
httpx
boto3